Extract serialization test fixtures
Move inline fixture construction from test_serialization.py into a
dedicated serialization_fixtures.py module to match the project
convention of separating fixtures from test logic.
